requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altgraph==0.10.2
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asn1crypto==0.24.0
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.1
bdist_mpkg==0.5.0
beautifulsoup4==4.11.1
bleach==4.1.0
build==1.2.2.post1
CacheControl==0.12.11
cachetools==5.5.1
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
cleo==1.0.0a5
click==8.1.3
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
crashtest==0.3.1
cryptography==37.0.4
cycler==0.12.1
dataclasses-json==0.6.7
datasketch==1.5.1
debugpy==1.8.12
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.6
distro==1.9.0
dulwich==0.20.46
durationpy==0.9
entrypoints==0.3
enum34==1.1.6
exceptiongroup==1.2.2
faiss-gpu==1.7.2
fastapi==0.115.7
fastjsonschema==2.16.1
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==25.1.24
fonttools==4.55.6
frozenlist==1.5.0
fsspec==2024.12.0
future==0.17.1
google-auth==2.38.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.4
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipaddress==1.0.22
ipykernel==6.4.1
ipysheet==0.5.0
ipython==7.28.0
ipython-blocking==0.3.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.0
jeepney==0.8.0
Jinja2==3.0.2
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyter-server==1.11.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==23.9.3
kiwisolver==1.4.7
kubernetes==32.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.2
langchain-text-splitters==0.3.5
langsmith==0.3.2
lockfile==0.12.2
macholib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.26.0
matplotlib==3.9.4
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
mmh3==5.1.0
modulegraph==0.10.4
monotonic==1.6
more-itertools==8.14.0
mpmath==1.3.0
msgpack==1.0.4
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==3.2.1
notebook==6.4.4
numpy==1.23.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.60.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
pipenv==2022.11.30
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
posthog==3.10.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
propcache==0.2.1
protobuf==5.29.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
py2app==0.7.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pylev==1.4.0
pynvml==12.0.0
pyparsing==3.0.9
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2021.3
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2024.11.6
requests==2.28.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
SecretStorage==3.3.3
Send2Trash==1.8.0
sentence-transformers==3.4.0
shellingham==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.16
starlette==0.45.3
sympy==1.13.1
tenacity==9.0.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.1.1
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.11.5
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.1.0
transformers==4.48.1
triton==3.1.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.12
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
watchfiles==1.0.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==14.2
Werkzeug==2.2.2
Whoosh==2.7.4
widgetsnbextension==3.5.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
dill
pyprobables
